[REMOTE-SHUFFLE-60] Support akey offset size bigger than 2GB

+++ shuffle-daos/src/main/java/org/apache/spark/shuffle/daos/DaosReaderBase.java
@@ -31,7 +31,7 @@ public abstract class DaosReaderBase implements DaosReader {
   protected Tuple2<String, Integer> curMapReduceId;
   protected Tuple2<String, Integer> lastMapReduceIdForSubmit;
   protected Tuple2<String, Integer> lastMapReduceIdForReturn;
-  protected int curOffset;
+  protected long curOffset;
   protected boolean nextMap;
 
   protected IODataDescBase currentDesc;
